Components as separate local package; Fix storybook webpack config and tsconfig;
