+++ packages/components/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@paritytech/components",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
