+++ packages/components/src/ButtonWithIcon.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent, ReactElement } from 'react';
 
-type ButtonProps = {
+export type ButtonProps = {
   label: string;
   Icon: React.FC<React.SVGProps<SVGSVGElement>>;
   onClick: (e?: MouseEvent<HTMLButtonElement, globalThis.MouseEvent>) => void | null;
