+++ packages/components/src/MainMenu/props.ts
