+++ packages/components/tsconfig.json
@@ -14,11 +14,9 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
-    "baseUrl": "./",
+    "baseUrl": "../",
     "paths": {
-      "~/*": ["./src/*"],
-      "@paritytech/components": ["../components/src"],
-      "@paritytech/components/*": ["../components/src/*"]
+      "~/*": ["./src/*"]
     }
   },
   "include": ["src", "jest", "__tests__"],
