+++ packages/playground/src/App/Header/index.tsx
@@ -1,31 +1,30 @@
-import { ReactElement, useState } from 'react';
+import { ReactElement, useReducer } from 'react';
 import { Logo } from '~/symbols';
-// import { MainMenu } from '~/components/MainMenu';
-import { MainMenu } from '~/App/Header/MainMenu';
+import { init, MainMenu, reducer } from '@paritytech/components/MainMenu';
 
 export const Header = (): ReactElement => {
-  const [dropdownOpen, setDropdownOpen] = useState<undefined | number>();
+  const [state, dispatch] = useReducer(reducer, init);
+
+  const items = [
+    { label: 'Apple', icon: '', onClick: () => alert('apple') },
+    { label: 'Pancake', icon: '', onClick: () => alert('pancake') },
+    {
+      label: 'Dishes',
+      icon: '',
+      subContent: () => <h2>More Dishes</h2>,
+    },
+    {
+      label: 'Wishes',
+      icon: '',
+      subContent: () => 'more wishes',
+    },
+  ];
 
   return (
     <div className="header">
       <Logo className="h-16 w-32" data-testid="headerLogo" />
       <div className="verticalDivider" />
-      <MainMenu />
-      {/*<MainMenu
-        openId={dropdownOpen}
-        items={[
-          { label: 'Compile', icon: '', onClick: () => {} },
-          { label: 'Share', icon: '', onClick: () => {} },
-          {
-            label: 'Settings',
-            icon: '',
-            sub: {
-              content: '',
-              trigger: () => setDropdownOpen(1),
-            },
-          },
-        ]}
-      />*/}
+      <MainMenu state={state} dispatch={dispatch} items={items} />
     </div>
   );
 };
