+++ packages/storybook/tsconfig.json
@@ -14,13 +14,12 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
-    "baseUrl": "../",
+    "baseUrl": "./",
     "paths": {
       "~/*": ["./src/*"],
-      "playground": ["./playground/src"],
-      "playground/*": ["./playground/src/*"]
+      "@paritytech/components": ["../components/src/"],
+      "@paritytech/components/*": ["../components/src/*"]
     }
   },
-  "include": ["src", "jest", "__tests__"],
-  "exclude": ["jest/svgTransform.ts"]
+  "include": ["stories"]
 }
